pyproject.toml
[project]
name = "e-commerce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.78",
    "boto3-stubs[athena,glue,s3]>=1.42.78",
    "dbt-athena-community>=1.10.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyathena>=3.30.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
